// src/app/models/wish-list.model.ts
export class WishList {
    _id: string = '';
    categoryId: string = '';
    productId: string = '';
    name: string = '';
    img: string = '';
    description: string = '';
    quantity: number = 0;
    price: number = 0;
    size: string = '';
    username: string = '';
    status: number = 0;
}